+++ global.d.ts
@@ -4,36 +4,25 @@ declare module 'nuomi-request' {
   export interface AxiosRequestOptions extends AxiosRequestConfig {
     extension?: string;
     cache?: boolean;
-    loading?: any;
     delay?: number;
   }
 
-  export interface CreateServicesFunction {
-    (urls: object, mockData?: any): object;
-  }
+  export const createServices: (urls: object, mockData?: any) => object;
 
-  export interface CreateMethodFunction {
-    (
-      name: string,
-      callback: (
-        url: string,
-        data?: object,
-        options?: AxiosRequestOptions,
-        ...rest: any[]
-      ) => AxiosPromise,
-      force?: boolean,
-    ): void;
-  }
+  export const createMethod: (
+    name: string,
+    callback: (
+      url: string,
+      data?: object,
+      options?: AxiosRequestOptions,
+      ...rest: any[]
+    ) => AxiosPromise,
+    force?: boolean,
+  ) => void;
 
-  export interface AxiosConfigFunction {
-    (options: object): void;
-  }
+  export const createMock: (options: object) => void;
 
-  export const createServices: CreateServicesFunction;
-
-  export const createMethod: CreateMethodFunction;
-
-  export const axiosConfig: AxiosConfigFunction;
+  export const axiosConfig: (options: object) => void;
 
   export const axios: AxiosStatic;
 }