+++ src/util.ts
@@ -1,8 +1,5 @@
-import { FormatResult } from './types';
-
-const PARAM_REGEXP = /(:\w+)/;
+const PARAM_REGEXP = /(\/:\w+)/;
 const EXT_REGEXP = /\.\w+$/;
-const SLASH_REGEXP = /\/+$/;
 
 export const globalWindow = typeof window !== 'undefined' ? window : global;
 
@@ -12,15 +9,22 @@ export const isObject = (obj: any) => {
 
 export const isString = (obj: any): boolean => typeof obj === 'string';
 
-export const format = (url: string, data: object): FormatResult => {
+/**
+ * @function 格式化url
+ * @param {string} url 请求url  /api /api/:id
+ * @param {object} data 请求参数
+ * @returns {string}
+ */
+export const formatURL = (url: string, data: object): string => {
   let newUrl = '';
 
   // 匹配动态参数
   if (PARAM_REGEXP.test(url)) {
+    const params = isObject(data) ? data : {};
     url.split(PARAM_REGEXP).forEach((path) => {
-      if (path.startsWith(':')) {
-        const key = path.substr(1);
-        newUrl += data[key] !== undefined ? data[key] : '';
+      if (path.startsWith('/:')) {
+        const key = path.substr(2);
+        newUrl += params[key] !== undefined ? `/${params[key]}` : '/';
       } else {
         newUrl += path;
       }
@@ -29,16 +33,20 @@ export const format = (url: string, data: object): FormatResult => {
     newUrl = url;
   }
 
-  return { url: newUrl, data };
+  return newUrl;
 };
 
+/**
+ * @function 组合url
+ * @param {string} url 请求url
+ * @param {string} extension 扩展符号, .do do .php php
+ * @param {boolean} cache 是否缓存，设置为false将会添加"?_=时间戳"参数
+ */
 export const combineURL = (url: string, extension?: string, cache?: boolean): string => {
   const paramStart = url.indexOf('?');
   const hasParam = paramStart !== -1;
   let path = hasParam ? url.substr(0, paramStart) : url;
   let search = hasParam ? url.substr(paramStart) : '';
-
-  path = path.replace(SLASH_REGEXP, '');
 
   if (extension && isString(extension) && !EXT_REGEXP.test(path)) {
     path += extension.startsWith('.') ? extension : `.${extension}`;